# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Grupo {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: string;
}

export interface Participante {
  id: number;
  nome: string;
  /** @nullable */
  chavePix: string | null;
  grupoId: number;
  criadoEm: string;
}

export interface GrupoComParticipantes {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: string;
  participantes: Participante[];
}

export interface ParticipanteInput {
  /** @minLength 1 */
  nome: string;
  /** @nullable */
  chavePix?: string | null;
}

export interface GrupoInput {
  /** @minLength 1 */
  nome: string;
  /** @minItems 1 */
  participantes: ParticipanteInput[];
}

export interface ParticipanteUpdate {
  /** @minLength 1 */
  nome?: string;
  /** @nullable */
  chavePix?: string | null;
}

export interface Divisao {
  participanteId: number;
  valorDevido: number;
}

export interface DespesaComDivisoes {
  id: number;
  descricao: string;
  valor: number;
  pagoPorId: number;
  grupoId: number;
  criadoEm: string;
  divisoes: Divisao[];
}

export interface DespesaInput {
  /** @minLength 1 */
  descricao: string;
  /** @minimum 0.01 */
  valor: number;
  pagoPorId: number;
  /** @minItems 1 */
  divisoes: Divisao[];
}

export interface DebitoItem {
  deId: number;
  paraId: number;
  valor: number;
}

export interface ParticipanteSaldo {
  participanteId: number;
  nome: string;
  saldoLiquido: number;
}

export interface Saldo {
  debitos: DebitoItem[];
  participantes: ParticipanteSaldo[];
  totalGasto: number;
}

export interface Pagamento {
  id: number;
  deId: number;
  paraId: number;
  grupoId: number;
  valor: number;
  criadoEm: string;
}

export interface PagamentoInput {
  deId: number;
  paraId: number;
  /** @minimum 0.01 */
  valor: number;
}