# lib/api-zod/src/generated/types/despesaComDivisoes.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { Divisao } from './divisao';

export interface DespesaComDivisoes {
  id: number;
  descricao: string;
  valor: number;
  pagoPorId: number;
  grupoId: number;
  criadoEm: Date;
  divisoes: Divisao[];
}