# lib/api-zod/src/generated/types/grupoComParticipantes.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { Participante } from './participante';

export interface GrupoComParticipantes {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: Date;
  participantes: Participante[];
}