# lib/api-zod/src/generated/types/pagamento.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */

export interface Pagamento {
  id: number;
  deId: number;
  paraId: number;
  grupoId: number;
  valor: number;
  criadoEm: Date;
}