# lib/api-zod/src/generated/types/pagamentoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */

export interface PagamentoInput {
  deId: number;
  paraId: number;
  /** @minimum 0.01 */
  valor: number;
}