# lib/api-zod/src/generated/types/participanteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rachador - Group expense splitting API
 * OpenAPI spec version: 0.1.0
 */

export interface ParticipanteInput {
  /** @minLength 1 */
  nome: string;
  /** @nullable */
  chavePix?: string | null;
}